Remove ACS support + refactoring
